refactor: 마이페이지(예약내역) api 분리
